[ISD-4095][Copilot] Remove repeated observers (#1062)
